LayoutTests/fast/sub-pixel/inline-block-with-padding.html

<!DOCTYPE html>
<html>
    <head>
        <style>
            a {
                padding: .3em 1.6em;
            }
        </style>
        <script src="../js/resources/js-test-pre.js"></script>
    </head>
    <body>
        <p>
            Both links should render the same and not wrap.
        </p>
        <div>
            <a style="display: inline-block;" href="#">This shouldn't wrap</a>
        </div>
        <div>
            <a href="#">This shouldn't wrap</a>
        </div>
        <script>
            var links = document.getElementsByTagName('a');
            var expectedWidth = links[1].offsetWidth;
            shouldBeCloseTo('links[0].offsetWidth', links[1].offsetWidth, 1);
        </script>
    </body>
</html>
